More robust read/write of file .i18n
We have received a few reports of broken ~/.i18n files
Try to make the code more robust, using "with" to open the files,
and remove redundancy in the code.
Only write the file when set the languages, and flush to
try avoid have a broken file.
When read the languages, get the values from the enviroment variables
set on bin/sugar. If the file ~/.i18n exists, was already read there.