Before open .i18n file to write, check if is really needed
When the user press the cancel button in the control panel section,
the ModelWrapper object (in gui.py) try to revert the changes
and write the initial values, even if nothing has changed.
As the recording of the values is triggered by a timeout,
we don't have another way to avoid race contions that verify
at the moment of write the file.