When add a new language set the right locale_code by default - Fixes #4878
When the user press the (+) button and a new language row is added,
the _add_row method receive locale_code=None. To make the method
manage this case properly, we need do two changes:
* Do not check in the available locales because None will not get any value.
* When add the locale_code to self._country_codes,
  set the locale based in the default values displayed to the user.

A consequence of the method not managing the None case,
id that changes will not be saved until the user change the default
value displayed to the user.