Refresh model in place for Journal ListView, fixes #4852
This commit changes the way the list view model updates occur.
Instead of replacing the ListModel, we now just change the result
set.  This means that the buggy scroll position setting code is no
longer needed.  Infact, a lot of code is removed.

However, the model is still replaced if the query is changed.  If the
query is changed, the length of the model could change and this would
effect the scroll bars.  It is eaiser to change the model in this
case, as we have no need to maintain the scroll position.