Make BundleRegistry monitor ATTRIBUTE_CHANGED GFileMonitor events
The BundleRegistry presented different issues which are related
to the use of GFileMonitor to detect bundle installation when being
done by external programs.

One example, as mentioned in the mailing list [1], is sugar-install-bundle,
where its execution time was drastically slowed down due to the changes added
in a previous commit [2].

The problem is that having SUGAR_PROFILE defined does not imply that the script
runs in the same runtime environment as the sugar shell. Therefore, each time
this script runs, it creates a completely new instance of the registry, which
is very expensive.

Another example, as mentioned in this ticket [3], where the registry misses to
install some bundles. This occurs because the GFileMonitor detects the bundle
root directory creation before the contents become available, e.g, it cannot find
the .info file. So the bundle gets ignored, never installed or never updated.

Due to problems like these we concluded [1] that we need an explicit way to tell
the registry to install or upgrade a bundle.

After experimenting with alternatives around DBus, I discarded all of them becuse
A) a session bus service will require the scripts to run in the same X session as
the sugar shell, which is clearly not useful, and B) a system bus endpoint would
require a completely new service and more config files, which I consider overkill
for just one method.

Instead, we could simply use one more GFileMonitor event to tell the BundleRegistry
about new or updated bundles.

This patch adds ATTRIBUTE_CHANGED event to the GFileMonitor callback, along with the
CREATED evenbt, so external scripts can trigger the bundle installation by doing just
one extra file system operation e.g, "touch /path/to/bundle", which is less expensive
and simpler than a DBus call. This simple solution could be use by any external script
installing bundles, from sugar-install-bundle to packaging post-install scripts.

[1] http://lists.sugarlabs.org/archive/sugar-devel/2015-August/050699.html
[2] 4b4b2fba1c37a9ad92ed30eb669b68552b62415
[3] https://bugs.sugarlabs.org/ticket/4841#comment:17

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>