fix busy cursor on logout
A previous patch to add busy cursor on logout never worked properly
because the cursor change was not flushed to the display.

Also, the busy cursor effect was not re-entrant, if the busy cursor was
already in use it would not be restored.

Add reference counted busy cursor methods, busy() and unbusy(), and
ensure the cursor change is flushed to the display.