add timeout in logout failure alert
Add a 30 second timeout for logout failure, so that logout is forced if
the user is inattentive.

TimeoutAlert does not support custom buttons, so the alert title and
message have been rewritten for Continue and Cancel buttons.

The response callbacks have been simplified to test only for Cancel.

Note: requires a patch to sugar3.graphics.alert.Alert in
sugar-toolkit-gtk3, otherwise a forced shutdown will occur 30 seconds
after the alert is shown, even if Cancel is clicked.