check before remove on control panel section end
A critical message appears in shell.log when a control panel section is
removed from screen:

(main.py:6713): Gtk-CRITICAL **: gtk_container_remove: assertion 'gtk_widget_get_parent (widget) == GTK_WIDGET (container) || GTK_IS_ASSISTANT (container)' failed

There is no other effect of the message.

Fix the warning by checking the object is in the container before trying
to remove it.

Since None can't be a child of the container, the check for None is
redundant, so is removed.