add quit failure alert
Add a quit failure alert to the menu and the control panel.  After 30
seconds with no click, continue with the quit.

Why?

The buddy menu has options for shutdown, restart or logout.  The control
panel may also restart.  These make Sugar quit.  Session manager sends a
save-yourself message to running activities, and waits for the
activities to respond.

When an activity does not respond, Sugar does not quit.  No reason is
given.

Adding the alert will allow user to cancel the shutdown, cancel the
control panel section change, and work the problem, or lose unsaved
work.

Caution: requires a patch to

    sugar-toolkit-gtk3:src/sugar3/graphics/alert.py

otherwise a forced shutdown will occur 30 seconds after the alert is
shown, even if Cancel is clicked.