Add zoom guesture for desktop views
Centeral the the nvaigation in Sugar is the zoom metaphor.  This
metaphor does not extend to the zoom guesture on multi touch
devices.

This commit adds a zoom guesture controller to the homewindow.
The user can then use the pinch to zoom guesture to switch between
the home, group and mesh views.