Claim file transfer channels, so empathy does not recieve them
If empathy is installed and the user is sent a file transfer (via
the journal), an empathy dialog will pop up asking the user where
to save the file (in the filesystem, not journal).  This conflicts
with the Sugar UI.

This is fixed using the channel approver.  In telepathyclient,
this commit adds file transfers to the list of channels that
the approver will inspect.  In filetransfer, this commit listens
for channels to approve and claims the file transfers for the
Sugar telepathy client.

Fixes #4907  <https://bugs.sugarlabs.org/ticket/4907>