background control panel, alpha level callbacks
Gio.Settings changes are memory and processor expensive, because the
instance persists until the change is flushed to disk.

Extra changes were happening during button creation, and for every click
of a button.  By default, the first button is created active by Gtk+.
When the code tries to set another button as active, the clicked
callback for the first button is triggered.  When followed by apply,
this can lead to incorrect setting.

Changes in this patch are to:

- ignore callbacks for a button changing from active to inactive, by
  checking get_active(),

- suppress callbacks during button creation, by connecting the toggled
  signal after the buttons have been created,

- simplify the button creation code.

Fixes #4912.