Allow activities to pass file transfers without shell notification
File transfer channels are a powerful solution for transfering data
1-to-1 over telepathy.  However, currently the shell will try to
claim all file transfer channels and show a notification to the
user.

This commit makes the shell ignore the channel if it has the
content type of 'x-sugar/from-activity'.  It will instead pass it
to the telepathy client specified in the file name.  The file name
was used to house the client name because there does not appear
to be a dedicated place for this information in telepathy.
Activities can still use the description to store their metadata.