Frame EventArea detects Gio.Settings changes
A rework of the Frame EventArea widget so that Gio.Settings changes are
applied immediately.

The EventArea widget is a collection of Gtk.Invisible widgets to detect
mouse movement into the frame area, in order to toggle the frame.

The 'changed' signal for the Gio.Settings schema now resizes and moves
the Gtk.Invisible widgets.  When a corner or edge delay is disabled, the
widgets are moved off-screen.  (Destroying and recreating the widgets
does not work.)

The widgets are enlarged by one pixel beyond the screen dimensions to
handle recent Gtk+ and X11 configurations that can position the cursor
outside the screen.  Fixes unresponsive frame gestures on commodity
hardware.  Underlying cause is floating-point mouse coordinates reported
in events.  Does not reproduce in sugar-build or virtual machines,
because mouse coordinates remain integer.