Add forward and back buttons in help view, fixes #4913
This allows the user to navigate, so that they do not get lost
if they click a hyperlink that takes them off the social help
website.  It also helps users feel safe if navigating the help
documentation, as they will not get lost.

This commit means relys on a get/set state implimentation
function, which is used when the user switches between modes.
This could be eaisly replaced with WebKit2 functions [1] when replaces the
time comes to migrate.

[1] http://webkitgtk.org/reference/webkit2gtk/unstable/WebKitWebView.html#webkit-web-view-get-session-state