Add proxy setting options in network control panel
Feature to set proxy settings for both manual and automatic (PAC) types
are added to the 'Network' section of 'My Settings' control panel. The
settings also support authentication for http proxy servers. The /jarabe/
main.py file where the proxy settings get exported finally is run after
any other standard scripts like /etc/environment and those inside
/etc/profile.d/, hence any proxy setting exported from such files gets
overwritten by the main.py. Therefore a new option ('Use system proxy')
has been added to the settings options list along with the standard
'none', 'manual', and 'auto', where in the main.py does not export the
 proxy settings to environment variables.

Once the settings are changed and the user clicks on toolbar to save it,
then the settings are verified. For manual proxy, the host address
that is provided is pinged and for auto proxy, the availability of the
pac file is checked. If the vefification fails, then an alert is shown
with options to either revert the changes or apply them anyway.

GObject signals are emitted by the view.py to perform the following:
- 'add-alert' : adds the restart alert so that it can be added only when
required.
- 'set-toolbar-sensitivity' : Set the sensitivity of top toolbar buttons.
The above two features were added to handle the gui so that there is a
straightforward user flow through the setting options.

Implementation:
It is based on a patch by Manuel Quiñones [1]. The GUI is based on the
Gnome 3 proxy settings to maintain consistency. The settings are stored
at two standard locations for diverse accessibility:
- schemas : custom defined schemas (eg 'org.sugarlabs.system.proxy')
- environment variables : (eg 'http_proxy')

Custom schemas (eg 'org.sugarlabs.system.proxy') are used based on
gnome schemas (eg 'org.gnome.system.proxy') [2]. It provides the
advantage of customizing the schemas. Hence the schema
'org.sugarlabs.system.proxy mode' was customized to include a new
mode 'system' corresponding to the system proxy option as above.

Six files are modified for this feature:
- /sugar/data/org.sugarlabs.gschema.xml : It defines the custom schemas
- /sugar/extensions/cpsection/network/view.py : It sets up the GUI and
stores the proxy settings to the schemas. A restart is required for the
proxy settings to take effect. Hence, this file also provides restart
options.
- /jarabe/main.py : It reads the proxy settings from schemas and exports
them to environment variables according to the options, when it is run
at the time of log in.
- /jarabe/controlpanel/gui.py : It handles the toolbar and restart alert.
Hence signal callbacks are added.
- /jarabe/controlpanel/sectionview.py : Added definition for the new
signals and a boolean 'show_restart_alert'.
- /sugar/extensions/cpsection/network/__init__.py : Add 'proxy'
keyword to the list to include it in search options.

Here is the related feature page [3] to be updated along with it.

[1] https://github.com/godiard/sugar/commit/ebfd1ed0f941dca982b90d0d08efda7c2ae75a84
[2] https://developer.gnome.org/ProxyConfiguration/
[3] https://wiki.sugarlabs.org/go/Features/Proxy_Settings