Connect to multiple XS server easily, fixes #362
Issue: As mentioned in the bug [1].
It is difficult to connect a Sugar device to multiple XS servers because
the user is required to perform few manual steps to clear a previous
registration. Also, new registration is performed even  when a laptop
tries to connect to an XS server, where it has already registered
previously. This caused the original backup-url and back-up path to
change on the server causing inefficiency.

Solution: The solution involves a few automated checks and tasks that
are performed to overwrite new registration details, store previous
registration details and restore the details is an old server is connected.
All GConf variables used by ds-backup are also updated with the server and
registration details. The identity of the server is added to known_hosts
file on successful registration to avoid ssh verification errors.

Server: I tried it on XSCE School Server release-6.0 on a CentOS7 machine.
No server side modification was done to archive it apart from enabling
'Services for XO Laptops such as Registration, Backup, and Activity Server'
in 'Configure' section of schoolserver admin. (http://schoolserver/admin/)
I used webpage available at schoolserver:5000 provided by xs-authserver,
where the list of registered laptops are present.

Here is a related feature page for details of this enhancement [2].

A link to the starting of a discussion thread about this feature. [3].

[1] https://bugs.sugarlabs.org/ticket/362
[2] https://wiki.sugarlabs.org/go/Features/Multi_XS-server_registration
[3] http://lists.sugarlabs.org/archive/sugar-devel/2016-April/051800.html