Display activity instance title in join requests
Currently, only the activity name is displayed.  This could be
confusing - "what is this write activity that you just invited me
to?  is it the kangaroo project, or the dolphin project?"

This commit displays the title as the palette secondary text, which
is consistent with how running activities are displayed in the frame.