Resume Activity Gsettings
A new boolean value is now added to gsettings to toggle the default start mode of an activity from home view. If the value is TRUE- Resumed instance opens
If the value is FALSE - New instance starts.
If the value is TRUE  - Activity resumes.

Terminal command to change the gSettings of this value is -
gsettings set org.sugarlabs.user resume-activity false/true