Language control panel: available locales search
When a system is installed with Ubuntu 16.04 and configured for English
language in the country Australia, the LANGUAGE list environment
variable is set to "en_AU:en", and there is no .i18n file.

Then when Sugar is used, the control panel shows "English/USA" twice
instead of "English/Australia" followed by "English/USA".  The control
panel lies.

Workaround is to change it and restart; this writes a new .i18n file and
the LANGUAGE variable is set to "en_AU.utf8:en_US.utf8" and the control
panel tells the truth.

Cause is that _available_locales has "en_AU.utf8" and so "en_AU" from
LANGUAGE is not found.

Fix is to extend the search of _available_locales to also match entries
ending in .utf8.