Language control panel: fix cancel changes growth
When a change is made to the list, and "Cancel changes" button is
pressed, the list of languages grows.

Cause is the list not cleared by undo before it is again loaded by
setup.

Fix is to clear the list by deleting all rows, and also avoid trying to
adjust the add/remove boxes in the previous row if there is no previous
row.