Language control panel: avoid deferred restart
When a change to language list is made, and the restart is deferred,
opening the list again shows old data, yet a restart through other means
will use the new data.

Regression was introduced by a84d6dd which removed the reading of .i18n
file, instead relying on environment variables.

The ambiguity is where the .i18n file has new data, but, while Sugar has
not restarted, the environment variables are old data.

Fix is to avoid the ambiguity by removing deferred restart.

An alternative fix is to add an explanation to the control panel for the
inconsistency between displayed language and language in use, but that
makes things more complicated.  There is very little need for deferring
restart on language change.