About my computer: fix section width.
On Ubuntu 16.04 with Sugar 0.109, and Ubuntu 14.04.2 or Fedora 20 with
Sugar 0.105.3, with a display of 1200x900 or 1366x768 pixels, clicking
on about my computer widens the control panel to the screen width.

This widening is unnecessary.

It also causes a brief flash of the control panel icons moved left.

All other control panel icons keep the control panel the same size.

On Fedora 18 with a display of 1200x900 pixels, there is no problem.

Fix is to specify a maximum text width, which is then ignored.

The mechanism for the fix is not understood.