Add wireless favourite locked emblem
Once connected to wireless network with encryption, the badge emblem
changes from a padlock to a star, and so a learner or teacher cannot
tell if encryption is being used.

Use a new favourite locked emblem when the wireless network is both a
favourite and supports encryption.

Also simplified and shortened the code by using a local variable.