Favourites view number unnecessary
On the home view, the label of the ring icon is "Favorites view 1", even
when there is only one view.  One view is the default.  There is no GUI
for changing the number of views.

Regression introduced by c5b0905916519927c35396a857bfe983b4912ea0.

Suppress the suffix when only one view is available.