Wireless right-to-forget #2264
Once a wireless network is connected to, Sugar persists in trying to
connect next time it starts.  Workaround is to "Discard wireless
connections" in network control panel.

Add a forget menu item to the wireless network icons.

--

After this patch was written, an alternate implementation was found in
https://bugs.sugarlabs.org/ticket/2264 written by Ezequiel Pereira
during Google Code-in 2014, that had not been merged:

https://bugs.sugarlabs.org/attachment/ticket/2264/0001-Added-Forget-option-to-network-menu.patch

Ezequiel's patch seems to go a bit further than necessary, and brings
the connection deletion into networkviews.py.  It may also need
rebasing.