Bind project list view to auto refresh
The _main_view_active boolean was refactored into an enum to better
accommodate how there is 3 views now (main, details and project).

Multi-selection was disabled in the project view.  It did not work
properly before this patch, due to the show_main_view call.

The show_main_view call from __selection_changed_cb was removed, as
this resulted in the main view being re-shown every refresh.  This call
did not appear to serve a purpose - selection changed would only be
called when the main view was focused (prior to adding the project
view).