About me - fix age picker spacing
When language is set to Spanish, the icons do not appear evenly spaced.

Each age is a Gtk.Grid containing an EventIcon and a Gtk.Label, but the
array of grids is held in a Gtk.Fixed with spacing set by _configure
method.  The design depends on the label fitting within the large icon
size of the style.

Wrap and centre the label.  This increases the chance that it will fit
within the large icon size.

Reported in https://bugs.sugarlabs.org/ticket/4967