Fix screen resize race
When an external monitor is connected or disconnected, about 25% of the
time the Sugar shell does not resize correctly; a size-changed signal
occurs but the workarea geometry has not changed.

Underlying problem is an update race in get_monitor_workarea, so the fix
is to call get_monitor_geometry for obtaining the width and height, but
continue to use the workarea offset for the later call to self.move().

Can be reproduced easily without an external monitor by switching
display panel resolution repeatedly;

    xrandr --output eDP1 --mode 1024x768
    xrandr --output eDP1 --mode 1366x768

Part of a fix for https://bugs.sugarlabs.org/ticket/4968