Fix radio off button
The control panel network radio button does nothing.  Occurs on Ubuntu
16.04 with NetworkManager 1.2.2.  Does not occur on Fedora 18.

Using the same API in a Python shell works fine.  The difference is that
the object is rapidly discarded within the function.

Retaining the NMClient object fixes the problem.  Perhaps the NM API has
changed to an asynchronous thread.