Revert "Refactor journal.objectChooser according to popup module in toolkit-gtk3"
This reverts a series of three commits;

* cadac88c1c880708bd2afe0e82a2c2356c4bd9df
* 0ba3171a3a5c92698a89553bdfa541c83e520694
* 739993b2c4e12faa33e6714a42e500c5d80fbb9f

On both Fedora 18 and Ubuntu 16.04, the series renders classroom laptops
useless in minutes or hours, because any activity that opens the object
chooser does fail to exit on quit, and processes leak and build-up on
the system until memory is depleted. Evidence of the problem is shown by
ps(1). Workaround is to reboot.

Reverting the series fixes the problem.

Handling the Wnck.Screen window-close signal in the same way as the
original code does not fix the problem.

Commit cadac88 doesn't say why the change was made, and doesn't appear
to add any functionality, so reverting seems the safest option.