Journal - fix batch continue button stall
A click on continue button will stall a batch operation.

Reproducer: create a reasonable number of journal objects so that erase
can be interrupted, select one then all journal objects, click on erase
and a message will show, click on continue and objects will begin to be
erased, click on continue again and progress will stall.

Hiding the button causes the Stop button to move unnecessarily, making
it harder to select.  Setting the button insensitive is difficult
because the toolkit does not expose it.

Ignore the button click if a batch operation is in progress.

Fixes #4971.  https://bugs.sugarlabs.org/ticket/4971