Revert "Fixed Issue: #4940 Journal entries name can be renamed to blank"
This reverts commit e05452a571ef9366b2e44d3c9c28d1b6182c4b50 as the
current behaviour is inconsistent across Sugar.

An journal entry title does not need to be non-blank; the learner should
be allowed to use a blank or empty title.  The title can always be
changed again.

The default title is never blank.  The title does not need to be unique;
it is not a file name.

Reference:
https://bugs.sugarlabs.org/ticket/4940