Save As - add GSettings and GConf
When an activity is to stop, and the save-as setting is enabled, and the
journal entry has not been named, show an alert asking for a name, with
keyboard focus in the name entry with the name selected.

Add the setting to both GSettings for GTK+ 3 activities, and GConf for
GTK+ 2 activities.

May be enabled from Terminal;

```
gsettings set org.sugarlabs.journal save-as true
gconftool-2 --type Boolean --set /desktop/sugar/journal/save_as true
```

https://wiki.sugarlabs.org/go/Features/Save-As

Signed-off-by: James Cameron <quozl@laptop.org>