Fix wording in deprecated controls specification
Marco Pesenti Gritti worked on the controls specification at OLPC with a team in 2007, but apart from some minor changes in 2010 and 2012 the document has not been changed.  It still contains GTK 2 elements and has not been maintained along with the other changes to the Sugar Toolkit for GTK 3.  The controls specification is of historical interest, otherwise we'd remove it.