[Final] Fix Journal entries can be renamed to blank via Detail View
Renaming items in listview now works as expected, an empty entry triggers a prompt
which asks if the user wants to leave the entry empty, clicking on Cancel or Okay discards
or saves the change.

This also fixes an existing issue in the journal where items weren't updated once they're
renamed and the user presses enter, this was caused by title-edit-finished callback in journalactivity
connecting key-press-event to __key_press_event_cb and that doesn't update the entry.

Signed-off-by: Ibiam Chihurumnaya <ibiamchihurumnaya@gmail.com>