Drag from clipboard to journal
Dragging text data from clipboard to journal now stores the data in the journal as expected,
also set timeout_id in drag-begin cb as the cb gets executed once a drag is started so it makes sense
to set the timeout_id there as it was set on drag-data-get cb and at the point it gets executed there's
already a call to remove it when a drag starts.

Switch to using Gtk.drag_finish in drag-data-received cb and encode text related data.

Free non-None URI data.

TODO
- Fix object recopying to the clipboard after it's dragged and saved to the journal.
- Test other types of data in the clipboard.

Signed-off-by: Chihurumnaya Ibiam <ibiam@sugarlabs.org>