Fix text clipping cannot be removed from Frame
The int datatype was used as an argument for the object-deleted signal, this meant that the Python
interpreter was responsible for deciding whether to use int32 or int64 based on the type of input
passed in.

Initially, the interpreter was selecting int32, but when larger input values were used, a TypeError
occurred because it could not convert int64 to int32 (gint).

Furthermore, the object_id was deleted when the object-deleted callback got executed leading to a
KeyError.

This commit changes the datatype of the function parameter to int64, ensuring that the
integer input is stored as int64 regardless of its range.