Remove use of Gtk.Settings and use Gio.
I've replaced every use of Gtk.Settings with Gio.Settings as in Wayland,
the settings are obtained from a settings portal or by reading the desktop
settings from Gio.

I've moved adding the sugar icons path to IconTheme into homewindow
as I changed from using the default as trying to get the default earlier seems
not to work as it returned None. I'll see if I need to revert this later.

Use Gdk.Seat for input device handling in cursortracker as in Gtk4,
input handling has been factored into Gdk.Seat and this prepares us for that.
I did this because I ran into issues when trying to get the default Gdk.Display,
and this was an attempt to fix it, I'm yet to test if it works as intended as I ran
into another issue which I'll be fixing next.

The issue is with initializing Gtk as in Wayland Gtk seems not to be initialized,
a possible fix for this is looking at using Gtk.Application as it doesn't
require you calling Gtk.init directly - which did me no good - as the Gtk.Application
startup signal handler calls it for you.

Signed-off-by: Chihurumnaya Ibiam <ibiam@sugarlabs.org>