Make Sugar shell a Gtk.Application
I decided to make the Sugar shell because it's the point of
registration for all running activities within Sugar, and it
makes sense to have this be the main application that keeps
track of windows to activity mapings which it already did.

I've also removed the use of SugarExt.wm* to get activity
related data like bundle_id and activity id and directly
set this data as attributes in the window itself in
sugar-toolkit-gtk3.

Wnck is no longer used as window management is now being
handled by Gtk.Application, so every Wnck related
functionality has been removed or replaced.

Every use of xid to identify windows have been removed and
replaced with bundle_id as that's now set on each window,
xids were how X got window IDs and we no longer need to do that.

My use of the shell in jarabe.main isn't pretty at the moment
and might change, that seems a great place to actually start
the shell before its use anywhere.

Some functions in jaraba.main are no longer needed and I've
removed them and moved any functionality that's needed to
other functions, there was a lot of checking which we don't
need to do anymore.

Signed-off-by: Chihurumnaya Ibiam <ibiam@sugarlabs.org>