Drop more use of GdkX11
I've replaced more use of GdkX11.X11Window with a Gdk.Window while
setting the foreign attribute on the window, I haven't passed a parent
because in most cases the default Gdk.Display is used and it automatically
becomes the parent if None is passed as the parent.

This also included removing any xid related use, although I've replaced
it with bundle_id, I'm yet to test if that's the best replacement and
will change it if I see it's not.

I've also removed the use of process_updates to force expose events to
be delivered immediately which is useful for updating the UI.

Signed-off-by: Chihurumnaya Ibiam <ibiam@sugarlabs.org>